refactor: 강의 데이터 적재 메서드 LectureService 이동
