+++ src/main/java/usw/suwiki/domain/lecture/service/LectureService.java
@@ -1,24 +1,33 @@
 package usw.suwiki.domain.lecture.service;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import usw.suwiki.domain.lecture.controller.dto.LectureDetailResponseDto;
-import usw.suwiki.domain.lecture.domain.repository.dao.LecturesAndCountDao;
-import usw.suwiki.domain.lecture.controller.dto.LectureResponseDto;
-import usw.suwiki.domain.lecture.controller.dto.LectureAndCountResponseForm;
-import usw.suwiki.domain.lecture.domain.Lecture;
-import usw.suwiki.domain.lecture.controller.dto.LectureFindOption;
-
+import java.io.FileReader;
+import java.io.IOException;
+import java.io.Reader;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+import lombok.RequiredArgsConstructor;
+import org.json.simple.JSONArray;
+import org.json.simple.JSONObject;
+import org.json.simple.parser.JSONParser;
+import org.json.simple.parser.ParseException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import usw.suwiki.domain.lecture.controller.dto.LectureAndCountResponseForm;
+import usw.suwiki.domain.lecture.controller.dto.LectureDetailResponseDto;
+import usw.suwiki.domain.lecture.controller.dto.LectureFindOption;
+import usw.suwiki.domain.lecture.controller.dto.LectureResponseDto;
+import usw.suwiki.domain.lecture.domain.Lecture;
+import usw.suwiki.domain.lecture.domain.repository.LectureRepository;
+import usw.suwiki.domain.lecture.domain.repository.dao.LecturesAndCountDao;
+import usw.suwiki.global.util.loadjson.JSONLectureVO;
 
 @Service
 @RequiredArgsConstructor
 public class LectureService {
 
     private final LectureCRUDService lectureCRUDService;
+    private final LectureRepository lectureRepository;
 
     @Transactional(readOnly = true)
     public LectureAndCountResponseForm readLectureByKeyword(String keyword, LectureFindOption option) {
@@ -40,6 +49,42 @@ public class LectureService {
     public LectureDetailResponseDto readLectureDetail(Long id) {
         Lecture lecture = lectureCRUDService.loadLectureFromId(id);
         return new LectureDetailResponseDto(lecture);
+    }
+
+    // TODO refactor: throws -> try catch
+    // TODO feat: 데이터 적재 과정을 admin API로 생성
+    @Transactional
+    public void bulkSaveJsonLectures(String filePath) throws IOException, ParseException {
+        Reader reader = new FileReader(filePath);
+        JSONParser parser = new JSONParser();
+        Object obj = parser.parse(reader);
+
+        JSONArray jsonArray = (JSONArray) obj;
+        if (jsonArray.size() > 0) {
+            for (int i = 0; i < jsonArray.size(); i++) {
+                JSONObject jsonObject = (JSONObject) jsonArray.get(i);
+                JSONLectureVO jsonLectureVO = new JSONLectureVO(jsonObject);
+
+                Optional<Lecture> optionalLecture = lectureRepository.findByExtraUniqueKey(
+                        jsonLectureVO.getLectureName(),
+                        jsonLectureVO.getProfessor(),
+                        jsonLectureVO.getMajorType()
+                );
+
+                if (optionalLecture.isPresent()) {
+                    Lecture lecture = optionalLecture.get();
+
+                    lecture.fixOmission(jsonLectureVO);
+                    lecture.addSemester(jsonLectureVO);
+
+                    lectureRepository.save(lecture);
+                } else {
+                    Lecture newLecture = jsonLectureVO.toEntity();
+                    lectureRepository.save(newLecture);
+                }
+
+            }
+        }
     }
 
     private LectureAndCountResponseForm readLectureByKeywordAndOption(String keyword, LectureFindOption option) {
